remove minimal benchmark.h
